refactored ai domains to functional pattern with domain-owned tools
- created withTool() helper for wrapping functions as AI tools
- refactored system-info, game-servers, apps, ops domains from classes to functions
- moved AI logic from routes/ai.ts to domains/ai/ structure
- each domain now owns its AI tools alongside business logic
- added AI tools: create/delete servers, app management, ops triggers
- added smart input focus to chat component
