+++ apps/nexus/src/domains/apps/index.ts
@@ -1,4 +1,4 @@
+export * from "./functions";
 export * from "./routes";
 export * from "./schema";
-export * from "./service";
 export * from "./types";
